+++ src/Corgi85/modules/CorgiModule.h
@@ -11,7 +11,7 @@ public:
   virtual void loop() = 0;
   virtual const char *name() = 0;
   virtual void cmd(String s) = 0;
-  virtual void raw(String s, uint16_t data_length)
+  virtual void raw(String s, uint32_t data_length)
   {
   }
 };
